Fix some more 'make check' warnings, ones that appear on non-Linux platforms.
git-svn-id: svn://svn.valgrind.org/valgrind/trunk@9238

+++ none/tests/rlimit_nofile.c
@@ -37,8 +37,9 @@ int main(int argc, char **argv)
 
    if (newrlim.rlim_cur != oldrlim.rlim_cur / 2)
    {
-      fprintf(stderr, "rlim_cur is %lu (should be %lu)\n",
-              newrlim.rlim_cur, oldrlim.rlim_cur / 2);
+      fprintf(stderr, "rlim_cur is %llu (should be %llu)\n",
+              (unsigned long long)newrlim.rlim_cur,
+              (unsigned long long)oldrlim.rlim_cur / 2);
    }
 
    if (newrlim.rlim_max != oldrlim.rlim_max)
